refactored tiledpositions
finished 1 big step in making ui work

+++ core/src/com/mygdx/game/Map/Map.java
@@ -16,11 +16,12 @@ import Utility.Utility;
 public abstract class Map implements AudioSubject{
     private static final String TAG = Map.class.getSimpleName();
 
+    //maybe remove final and make this automatic based on map + change name
     public final static float UNIT_SCALE  = 1/32f;
     public final static int TILE_HEIGHT  = 32;
     public final static int TILE_WIDTH  = 32;
-    public static int TILE_HEIGHT_PIXEL  = 32;
-    public static int TILE_WIDTH_PIXEL  = 32;
+    public static float TILE_HEIGHT_PIXEL  = 32.0f;
+    public static float TILE_WIDTH_PIXEL  = 32.0f;
     
     private Array<AudioObserver> _observers;
 
@@ -88,8 +89,8 @@ public abstract class Map implements AudioSubject{
     }
     
     public void updatePixelDimensions() {
-		TILE_WIDTH_PIXEL = Gdx.graphics.getWidth() / mapWidth;
-		TILE_HEIGHT_PIXEL = Gdx.graphics.getHeight() / mapHeight;
+		TILE_WIDTH_PIXEL = Gdx.graphics.getWidth() / (float) mapWidth;
+		TILE_HEIGHT_PIXEL = Gdx.graphics.getHeight() / (float) mapHeight;
     }
     
     private void disposeMapAndStage() {
